migration/requirements.txt
cryptography==43.0.3
mariadb==1.1.10
python-dotenv==1.0.1
tqdm==4.66.6
